Resource Based Authorization

+++ DineHub.Application/Commands/Restaurants/DeleteRestaurantCommandHandler.cs
@@ -1,12 +1,14 @@
 using DineHub.Domain.Entities;
+using DineHub.Domain.Enums;
 using DineHub.Domain.Exceptions;
+using DineHub.Domain.Interfaces;
 using DineHub.Domain.RepositoryContracts;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
 namespace DineHub.Application.Commands.Restaurants;
 
-public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantRepository restaurantRepository) : IRequestHandler<DeleteRestaurantCommand>
+public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantRepository restaurantRepository, IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +19,9 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         if (restaurant == null)
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
 
+        if (!restaurantAuthorizationService.Authorize(restaurant, RecourseOperation.Delete))
+            throw new ForbiddenException("This operation is forbidden for you");
+        
         await restaurantRepository.DeleteRestaurantAsync(restaurant);
     }
 }
